backend/requirements.txt
# # Core framework
# Flask==3.0.2
# SQLAlchemy==2.0.29
# pydub==0.25.1
# pillow==11.2.1
# selenium==4.31.0
# selenium-wire==5.1.0
# undetected_chromedriver==3.5.5
# packaging>=23.1
# blinker==1.7.0
# setuptools==78.1.0

# PyJWT
# cryptography
# clerk-backend-api

# pathlib
# requests
# flask_cors
# flask
# regex

# Appium-Python-Client
# python-dotenv
# ffmpeg-python
# docker
# gunicorn
# psycopg2-binary
# pandas
# aiortc
# av
# deep-translator
# psutil

# Flask-SQLAlchemy
# Flask-JWT-Extended
# alembic

aioice==0.10.2
aiortc==1.13.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
appium-python-client==5.2.7
attrs==25.4.0
av==14.2.0
beautifulsoup4==4.14.3
blinker==1.7.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
clerk-backend-api==4.2.0
click==8.1.8
cryptography==45.0.7
deep-translator==1.11.4
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.3.1
ffmpeg-python==0.2.0
flask==3.0.3
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flask-sqlalchemy==3.1.1
future==1.0.0
google-crc32c==1.8.0
greenlet==3.2.5
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ifaddr==0.2.0
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
kaitaistruct==0.11
mako==1.3.10
markupsafe==3.0.3
mercurial==5.6.1
numpy==2.0.2
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pathlib==1.0.1
pillow==11.2.1
psutil==7.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pyee==13.0.1
pyjwt==2.11.0
pylibsrtp==0.12.0
pyopenssl==25.3.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
regex==2024.11.6
requests==2.32.3
selenium==4.31.0
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlalchemy==2.0.48
trio==0.31.0
trio-websocket==0.12.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
undetected-chromedriver==3.5.5
urllib3==2.6.3
websocket-client==1.9.0
websockets==15.0.1
werkzeug==3.1.6
wsproto==1.2.0
zipp==3.23.0
zstandard==0.25.0
rich
scikit-image  
imagehash 
